Formatting AND change pageable queries to use int instead of long

+++ control-admin/src/main/java/fi/nls/oskari/control/admin/UsersHandler.java
@@ -47,8 +47,8 @@ public class UsersHandler extends RestActionHandler {
     public void handleGet(ActionParameters params) throws ActionException {
         final JSONObject response;
         long id = getId(params);
-        long limit = params.getHttpParam(PARAM_LIMIT, 0);
-        long offset = params.getHttpParam(PARAM_OFFSET, 0);
+        int limit = params.getHttpParam(PARAM_LIMIT, 0);
+        int offset = params.getHttpParam(PARAM_OFFSET, 0);
         String search = params.getHttpParam(PARAM_SEARCH);
         try {
             if (id > -1) {
